Fixed a problem with infinite scrolling not working.
Previously, our JS was loaded in the head. JS gets executed as its
loaded, so that meant that sometimes, the JS would execute before the
page was fully loaded. The pagination element, which the infinite scroll
JS depends on, wouldn't be there when the JS executed, causing the JS to
quietly fail. This was a particular problem after creating a new post or
forum thread because this would invalidate the cached copy of the forum
threads (because they are ordered on updated_last, which is updated on
post creation), causing the application to request the forum threads
from the database, delaying the end of the page almost always until
after the JS was loaded breaking infinite scrolling. It also sometimes
broke for no explicit reasons, it was just feeling slow that moment.
Moving the JS load code to the bottom of the layout fixed this problem,
and any other similar problems.